Added JsonIgnoreProperties to Task to fix new (undocumented) Asana API Properties
The Asana API now sends an additional property “due_at” (seems to be a
duplicate of “due_on”). Observed the first time 2015-01-09. Mail to API
support sent.

Added annotation to ignore unknown properties. Could be a good idea in
general in order to not break the code when new properties are added to
the Asana API. This would however also mean that changing property
names will not lead to an Exception but to unexpected null values.